+++ .github/workflows/any_changes.yaml
@@ -26,8 +26,6 @@ jobs:
         
         - name: Test documentation builds
           run: make documentation
-          env:
-            HUGGING_FACE_TOKEN: ${{ secrets.HUGGING_FACE_TOKEN }}
 
         - name: Check documentation build
           run: |
