+++ .github/workflows/code_changes.yaml
@@ -44,5 +44,3 @@ jobs:
           
           - name: Run tests
             run: make test
-            env:
-              HUGGING_FACE_TOKEN: ${{ secrets.HUGGING_FACE_TOKEN }}
