+++ .github/workflows/publish_documentation.yaml
@@ -29,8 +29,6 @@ jobs:
         
       - name: Build documentation
         run: make documentation
-        env:
-          HUGGING_FACE_TOKEN: ${{ secrets.HUGGING_FACE_TOKEN }}
       
       - name: Deploy documentation
         uses: JamesIves/github-pages-deploy-action@releases/v3
