+++ policyengine/utils/data/caching_google_storage_client.py
@@ -21,20 +21,6 @@ class CachingGoogleStorageClient(AbstractContextManager):
     def _data_key(self, bucket: str, key: str, version: str | None) -> str:
         return f"{bucket}.{key}.{version}.data"
 
-    def _get_latest_version(self, bucket: str, key: str) -> str | None:
-        """
-        Get the latest version of a blob in the specified bucket and key.
-        If no version is specified, return None.
-        """
-        blob = self.client.client.get_bucket(bucket).get_blob(key)
-        if blob.metadata is None:
-            logging.warning(
-                "No metadata found for blob, so it has no version attached."
-            )
-            return None
-        else:
-            return blob.metadata.get("version")
-
     # To absolutely 100% avoid any possible issue with file corruption or thread contention
     # always replace the current target file with whatever we have cached as an atomic write.
     def download(
@@ -45,7 +31,7 @@ class CachingGoogleStorageClient(AbstractContextManager):
         """
         if version is None:
             # If no version is specified, get the latest version from the cache
-            version = self._get_latest_version(bucket, key)
+            version = self.client._get_latest_version(bucket, key)
             logging.warning(
                 f"No version specified for {bucket}, {key}. Using latest version: {version}"
             )
